+++ backend/alembic/versions/496bd2f8c36b_initial_migration.py
@@ -1,0 +1,44 @@
+"""Initial migration
+
+Revision ID: 496bd2f8c36b
+Revises: 
+Create Date: 2025-04-19 11:05:27.530989
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '496bd2f8c36b'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('appointments',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('first_name', sa.String(length=30), server_default='John', nullable=False),
+    sa.Column('last_name', sa.String(length=30), server_default='Doe', nullable=False),
+    sa.Column('start_time', sa.DateTime(), nullable=False),
+    sa.Column('duration', sa.Integer(), nullable=False),
+    sa.Column('acuity_created_at', sa.DateTime(), nullable=False),
+    sa.Column('acuity_deleted_at', sa.DateTime(), nullable=True),
+    sa.Column('is_deleted', sa.Boolean(), nullable=True),
+    sa.Column('created_at_here', sa.DateTime(), server_default=sa.text('(CURRENT_TIMESTAMP)'), nullable=False),
+    sa.Column('last_modified_here', sa.DateTime(), server_default=sa.text('(CURRENT_TIMESTAMP)'), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('appointments')
+    # ### end Alembic commands ###
