--- backend/alembic/versions/6c19de6a9f8b_add_start_time_and_duration_columns.py
@@ -1,37 +1,0 @@
-"""add start_time and duration columns
-
-Revision ID: 6c19de6a9f8b
-Revises: 9a56ba9c0e75
-Create Date: 2025-04-19 00:39:48.935675
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '6c19de6a9f8b'
-down_revision: Union[str, None] = '9a56ba9c0e75'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table("appointments") as batch_op:
-        batch_op.add_column(sa.Column('start_time', sa.DateTime(), nullable=False, 
-                            server_default=sa.text("'2023-01-01 00:00:00'")))
-        batch_op.add_column(sa.Column('duration', sa.Integer(), nullable=False, 
-                            server_default=sa.text("60")))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('appointments', 'duration')
-    op.drop_column('appointments', 'start_time')
-    # ### end Alembic commands ###
