--- backend/alembic/versions/e90e0db5c858_change_name_to_first_and_last_name.py
@@ -1,37 +1,0 @@
-"""change name to first and last name 
-
-Revision ID: e90e0db5c858
-Revises: 6c19de6a9f8b
-Create Date: 2025-04-19 01:37:53.706351
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'e90e0db5c858'
-down_revision: Union[str, None] = '6c19de6a9f8b'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table('appointments') as batch_op:
-        batch_op.add_column(sa.Column('first_name', sa.String(length=30), nullable=False, server_default=sa.text("John")))
-        batch_op.add_column(sa.Column('last_name', sa.String(length=30), nullable=False, server_default=sa.text("Doe")))
-        batch_op.drop_column('name')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('appointments', sa.Column('name', sa.VARCHAR(length=30), nullable=False))
-    op.drop_column('appointments', 'last_name')
-    op.drop_column('appointments', 'first_name')
-    # ### end Alembic commands ###
